Add `build_npm` script to generate node compatible code
The compatibility is created by using `dnt` deno to node package. The
output is a distributable with regular javascript and typescript types
that work in both esm and cjs module resolutions.

The build command is invoked by calling `deno task build-npm --mod
<mod_name>` or by directly using `deno run ./build_npm.ts`.

In order for the build_npm to work a module name needs to be provided.
Module name in this instance can be a directory with module file
`mod.ts` and a `pkg.json` file. `pkg.json` file is used to template some
of the package.json fields that are not generated, but still wanted in
the final distribution.

The package version is provided with a flag `--version` and without the
version flag the build script will create a prerelease with
0.0.0-<commit_sha> as the version.