Add `Option` utility type
Option represents the presence of a value `Some` or the absence of a
value `None`. It's commonly utilized to manage computations that might
or might not yield a result, or those that could potentially fail.

The following functions are provided by the option type:

Option.bind - Applies a function `fn` to the content of option `T` and transforms
it into an option `U`.

Option.map - Applies a function `fn` to option value `T` and transforms it into
value `U`.

Option.match - Transforms the option value `T` into `U` using `onSome` and then
returns `U`. If the option is None, it uses `onNone` and returns `U`.

Option.unwrap - Returns the value `T` from the associated option if it is `Some`;
otherwise it will throw.

Option.unwrapOr - Returns the value `T` from the associated option or returns the
default value `U` if the option is `None`.

Option.isSome - Returns `true` if the option is `Some`

Option.isNone - Returns `true` if the option is `None`