Use `flatMap` instead of `bind`
The `bind` as a term is already used as a function method in ECMAScript
and conveys a different meaning (compared to F# bind where the term used
in the Option API was originated from). Using `flatMap` as the method
name for Option API instead of `bind` makes the API more consistent and
frankly more idiomatic. For example the term `flatMap` is already found
as an `Array` method in ECMAScript. [ECMAScript
specification](https://tc39.es/ecma262/multipage/indexed-collections.html#sec-array.prototype.flatmap).