Add `Result` utility type
Result encapsulates two possible values: a successful result `Ok` or an error
`Err`. Result allows to chain together a series of operations that can
potentially fail and propagate the error through the computations. This enables
to write code that is more focused on the 'happy path' and separate the error
handling logic in a clean and concise way.

Result is similar to Option, the main difference is that it holds either a
value or an error, and not value or none.

The following functions are provided by the option type:

Result.flatMap - Applies a function `fn` to the content of a result `T` and
transforms it into a result containing value `U`.

Result.inspect - Calls the provided function `fn` with a reference to the
contained result value `T` if the result is ok.

Result.inspectErr - Calls the provided function `fn` with a reference to the
contained result error value `E` if the result is err.

Result.map - Applies a function `fn` to result value `T` and transforms it into
value `U`.

Result.mapErr - Applies a function `fn` to result error value `E` and
transforms it into value `F`.

Result.match - Transforms the result value `T` into `U` using `onOk` and then
returns `U`. If the result is Err, the error value `E` is transformed to `U`
with `onErr` and then returns `U`.

Result.unwrap - Returns the value `T` from the associated result if it is `Ok`;
otherwise it will throw.

Result.unwrapErr - Returns the value `E` from the associated result if it is
`Err`; otherwise it will throw.

Result.unwrapOr - Returns the value `T` from the associated result or returns the
default value `U` if the result is `Err`.

Result.isOk - Returns `true` if the result is `Ok`.

Result.isErr - Returns `true` if the result is `Err`.