Unify `Result` type function signatures and documentation
Previously either Ok or Err had different function signatures for same
method names. These methods might take parameters in the other class and
not in the other and vice versa. For example map takes parameters as a
method for Ok and did not take parameters when defined as a method for
Err.

By having the function signatures follow exactly how the type is defined
makes the API more concise and have less confusing parts.

Documentation now reflects the change of function signatures for Ok and
Err and any mentions of "Impossible state" has been removed.