Make `unwrapOr` default value use contained type `T`
Previously the `unwrapOr` would take a generic `U` as the type for the
default value. This was a bit inconsistent since one could return a
different type than what is contained inside Option or Result.

To make this more consistent the unwrapOr now forces the caller to pass
a value that is the same type as the contained value `T`.