Fix `unwrapOr` on None and Err
The previous commit 31f0cc2 made `unwrapOr` act more consistently with
the contained value `T`, however, with None and Err the contained value
is `never`, since it does not exist. In these cases the `unwrapOr`
defaultValue would be forced to be given as `never` type, which is not
correct. The type signatures were adjusted accordingly so the correct
type `U` is inferred from the defaultValue.