Add first benchmarks for Option and Result types
The benchmarks are done against the examples with fetch and json parse.
Keep in mind that this is only preliminary benchmarks to have an idea of
the performance hit or gain when using this library. The option and
result benchmarks are run against the native implementations (without
the library functions).

The historic data can be used to measure if performance has improved or
degraded while making changes to the code.