Fix type inference for match and flatMap in Result
Previously match was strict of it's return type and forced the caller to
return the same return type in both callbacks onOk and onErr. This
changes the match function to allow a more loose type to be returned.
Both callbacks are now isolated from their return type (U1, U2).

FlatMap type inference could fail to provide a proper type. More
explicit overload for the flatMap allows for better type inference. This
type inference problem was prominent when one would return Ok<void> |
Ok<SomeType> | Err<string> union from a function as the return type,
which, previously would just fail to a type inference problem.

Unwrap did not work properly due to Err unwrap returning never type.
This was fixed with a type magic by making the unwrap generic, but
forcing it to be inhereted from the original T type.