Add Option.from function
`Option.from` function merges the functionality of fromPromise and
fromNullable to a one function. The function return type is a recursive
type, which means that it can infer from the given value what the actual
return type should be and thus matches with the inner logic of the
function and returns the expected type. The recursive type is called
`From`. The function itself streamlines the conversion to Option and
allows to use just one function when converting values to Options.

```ts
// Easily convert optional values into options
opt = Option.from(maybeString);

// Or even functions!
opt = Option.from(() => maybeString);

// Or promises
opt = Option.from(Promise.resolve(42));
```