Add Result.from function
`Result.from` function merges the functionality of fromPromise and
fromThrowable to a one function. The function return type is a recursive
type, which means that it can infer from the given value what the actual
return type should be and thus matches with the inner logic of the
function and returns the expected type. The recursive type is called
`From`. The function itself streamlines the conversion to Result and
allows to use just one function when converting values to results.

The Result.from function accepts an optional second parameter
`expected`, which if given will replace any errors thrown or rejected
and be returned instead.

```ts
// Easily convert async functions to results
result = Result.from(fetch("https://some-address.com/api/get_endpoint"));

// Or promises
result = Result.from(Promise.resolve(42));

// Or primitive values
result = Result.from(42);
```