Add github-actions workflow for publishing automatically
The verify workflow was merged into main-workflow that also handles the
publishing part of the modules.

The modules will be published only on these conditions:

1. branch is main
2. deno.json under the module folder contains version field changes

The publishing happens automatically to jsr.io by calling npx jsr
publish in either of the modules directory depending on if the
conditions were met.