Improve return type correctness for Option.from function
Previously the from function could return also Some explicitly, this was
changed to always return Option<T> or None in obvious cases. A never
type is returned if `T extends undefined`, which means that we return a
"None" type from the function. In this case we skip the explicit None
type in the return type and only return type Option<T>, which implicitly
includes a None type.