Use `undefined` instead of `void` as the unit type for Result
Reason for this is that the `void` type and `undefined` type behave
differently. `undefined` is actually the correct type to use here as the
unit type as `void` type is explicitly used for functions that return
nothing. In runtime void type is undefined.

Type inference behaves differently with undefined and void types.
Undefined is more lenient and works better when performing type
inference.