Add actual implementation for Result.from expection map function
Previously the mapper function was only taken into account on the type
level and was missing the implementation. Fixed the error type to
correctly be inferred as the return type of the given mapper function.
Explicitly set the mapper value as `any` to indicate that it is not
type-safe to use.