Option module overhaul
BREAKING CHANGE

First the big change for the API! Previously users needed to create Some
type and None type through the namespace function `.some` or `.none`
respectively. This felt a bit tedious at times as calling
`Option.some(42)` is quite verbose. In this overhaul `Some()` and
`None()` functions were introduced. These are actually behind a Proxy
that will apply the arguments to the class constructors. None being a
static entity. To achieve the API we use a Proxy to trap the function
call and then return the constructed entity. The static functions are
still kept for more concise operability with callbacks and to retain
backwards compatibility. The breaking change is when callers need to
type with either Option.Some or Option.None. These will now result in
compiler error and should be changed to just Some or None.

Other big change introduced in this commit was the use of abstract class
and static methods for the Option type. This allows for better
interoperability and for example the documentation can now be written
only for the abstract methods and these will still be visible from the
Some and None classes.

README was also updated to encompass these changes. Static methods and
higher-order functions were documented in the README usage section as
well.

Tests and examples were updated to match the new API.