Result module overhaul
BREAKING CHANGE

First the big change for the API! Previously users needed to create Ok
and Err types through the namespace function `.ok` or `.err`
respectively. This felt a bit tedious at times as calling
`Result.ok(42)` is quite verbose. In this overhaul `Ok()` and `Err()`
functions were introduced. These are actually just the same static
methods as previously, but acting as constructor functions similar to
JavaScript `String` and `Number` constructors. The static functions are
still kept for more concise operability with callbacks and to retain
backwards compatibility. The breaking change is when callers have types
Result.Ok or Result.Err. After these changes these types will now result
in compiler error and should be changed to Ok or Err instead.

Other big change introduced in this commit was the use of abstract class
and static methods for the Result type. This allows for better
interoperability and for example the documentation can now be written
only for the abstract methods and these will be inherited by extending
and appear visible from the Ok and Err class methods.

README was also updated to encompass these changes. Static methods,
constructor functions and higher-order functions are documented in the
README usage section.

Tests and examples were updated to match the new API.