Add `expect` and `expectErr` for Result
Result.expect will return the wrapped value if the result is Ok. If the
result is Err, the function will throw an error with the given input
value as the error message. Similarly, Result.expectErr will return the
wrapped value if the result is Err, and throw an error if the result is
Ok.