Use matrix strategy in main workflow to reduce repetition
Release and publish jobs had a bit of duplication due to two different
modules (option, result) and registries (npm, jsr). This is now handled
with a two matrices. First one determines if a particular module should
be published and the second one will run the publish command for the
approved modules.