Add `Option.all` and `Option.any` static methods
Option.all returns all Some option values as an array within a Some
option. If None option exists in the array, None is returned. An empty
array signifies success, resulting in a Some with an empty array.

Option.any returns the first Some option encountered. If no Some options
are found in the array None is returned.