Add `Result.all` and `Result.any` static methods
Result.all returns all Ok result values as an array within an Ok result
if no Err results are present. If any Err result exists in the array,
the first one is returned. An empty array signifies success, resulting
in an Ok with an empty array.

Result.any returns the first Ok result encountered. If no Ok results are
found in the array, all Err result values are returned as an array
within an Err result.