Add `terser` to minify code published to npm
Terser is a JavaScript mangler/compressor. More info here:
https://github.com/terser/terser

By utilizing terser the size of the output distribution files is
considerable smaller as the code is compressed and mangled to minimum
functional state. Terser also removes comments, which in turn replaces
the naive comment remove solution that was relying on regex.

The compressed output file size is around 50% smaller from 4.33kb to
2.2kb for option module.