Add doc testing for Option and Result modules
Doc testing is really useful to keep the example documentation up to
date, for example in a case of API change in some particular function.
The example code is also statically checked by the type checker.

Add import statements to examples, as otherwise these examples would not
pass the test as the called functions would not be found in the scope.

Remove NonNullable requirement from Some, this made the returned types
wider e.g. literal number `42` type turned into a `number` type.